requirements.txt
fastapi
uvicorn
pydantic

streamlit
requests

sentence-transformers
faiss-cpu
rank-bm25

numpy
scikit-learn

tqdm

python-multipart

ollama